emotion_adaptive_study_assistant/requirements.txt
# Emotion-Adaptive Study Assistant - Dependencies
# ================================================

# Backend Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Frontend Framework
flask==3.0.0
flask-cors==4.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Computer Vision & Facial Emotion Detection
opencv-python==4.8.1.78
deepface==0.0.79
tf-keras==2.15.0

# Audio Processing & Voice Sentiment
librosa==0.10.1
sounddevice==0.4.6
numpy==1.24.3
scipy==1.11.4
numba==0.58.1

# Machine Learning
scikit-learn==1.3.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.2

